[MOVE]/#131 패킹리스트 초대 URI 변경
