+++ src/controllers/ListController.ts
@@ -8,35 +8,6 @@ import { ListService } from '../services';
 import db from '../loaders/db';
 
 /**
- *  @route GET /invite/:inviteCode
- *  @desc invite list
- *  @access private
- **/
-const inviteList = async (req: Request, res: Response) => {
-  let client;
-
-  const inviteCode = req.params.inviteCode;
-  try {
-    client = await db.connect(req);
-    const userId = req.body.user.id;
-    const data = await ListService.getPackingByInviteCode(client, inviteCode, userId);
-    if (data === 'no_list') {
-      res.status(statusCode.NOT_FOUND).send(util.fail(statusCode.NOT_FOUND, message.NO_LIST));
-    } else {
-      res
-        .status(statusCode.OK)
-        .send(util.success(statusCode.OK, message.SUCCESS_INVITE_TOGETHER_PACKING, data));
-    }
-  } catch (error) {
-    res
-      .status(statusCode.INTERNAL_SERVER_ERROR)
-      .send(util.fail(statusCode.INTERNAL_SERVER_ERROR, message.INTERNAL_SERVER_ERROR));
-  } finally {
-    if (client !== undefined) client.release();
-  }
-};
-
-/**
  *  @route PATCH /list/title
  *  @desc update list title
  *  @access private
@@ -193,7 +164,6 @@ const getSharedList = async (req: Request, res: Response) => {
 };
 
 export default {
-  inviteList,
   updateDate,
   updateTitle,
   updateMyTemplate,
