+++ src/routes/ListRouter.ts
@@ -7,8 +7,6 @@ const router = Router();
 
 router.get('/:listType/share/:inviteCode', ListController.getSharedList);
 
-router.get('/invite/:inviteCode', auth, ListController.inviteList);
-
 router.patch(
   '/title',
   [body('id').notEmpty(), body('title').notEmpty(), body('isAloned').notEmpty()],
