+++ src/services/ListService.ts
@@ -9,68 +9,6 @@ import { SharedTogetherListResponseDto } from '../interfaces/ITogetherList';
 import { aloneCategoryResponse } from '../modules/aloneCategoryResponse';
 import { togetherCategoryResponse } from '../modules/togetherCategoryResponse';
 
-const getPackingByInviteCode = async (
-  client: any,
-  inviteCode: string,
-  userId: number,
-): Promise<ListInviteResponseDto | string> => {
-  try {
-    const { rows: packingList } = await client.query(
-      `
-      SELECT tapl.id::text, t.group_id, t.id AS "togetherId"
-      FROM "together_packing_list" as t
-      JOIN "packing_list" as pl ON pl.id = t.id
-      JOIN together_alone_packing_list tapl on t.id = tapl.together_packing_list_id
-      WHERE t.invite_code = $1 AND pl.is_deleted = false
-      `,
-      [inviteCode],
-    );
-    if (packingList.length === 0) return 'no_list';
-
-    // 이미 추가된 멤버인지
-    let isMember = false;
-
-    const { rows: existMember } = await client.query(
-      `
-          SELECT *
-          FROM "user_group" as ug
-          WHERE ug.user_id = $1 AND ug.group_id = $2
-        `,
-      [userId, packingList[0].group_id],
-    );
-    if (existMember.length > 0) isMember = true;
-
-    if (isMember === true) {
-      const { rows: newPackingList } = await client.query(
-        `
-            SELECT tal.id::text
-            FROM "together_alone_packing_list" tal 
-            JOIN "folder_packing_list" fl ON tal.my_packing_list_id = fl.list_id
-            JOIN "folder" f ON f.id = fl.folder_id
-            JOIN "packing_list" pl ON pl.id =  fl.list_id
-            WHERE tal.together_packing_list_id = $1 AND f.user_id = $2 AND pl.is_deleted = false
-          `,
-        [packingList[0].togetherId, userId],
-      );
-      if (newPackingList.length === 0) return 'no_list';
-      const data: ListInviteResponseDto = {
-        id: newPackingList[0].id,
-        isMember: isMember,
-      };
-      return data;
-    }
-
-    const data: ListInviteResponseDto = {
-      id: packingList[0].id,
-      isMember: isMember,
-    };
-    return data;
-  } catch (error) {
-    console.log(error);
-    throw error;
-  }
-};
-
 const updateTitle = async (
   client: any,
   titleUpdateDto: TitleUpdateDto,
@@ -421,7 +359,6 @@ const getSharedList = async (
 };
 
 export default {
-  getPackingByInviteCode,
   updateTitle,
   updateDate,
   updateMyTemplate,
